+++ views/engagement-referrers.html
@@ -1,0 +1,9 @@
+<h1>Articles Read</h1>
+
+<p class="o-techdocs-leadbody">
+	External referrers for our two types of users (daily read < and >= 1.5)
+	<br />
+	<small>This report can take a while to run, so please be patient</small>
+</p>
+
+<section id="graph__container"></section>
