+++ views/pages/gallery.html
@@ -131,6 +131,7 @@
 	</div>
 	<div class="o-aside-panel__body">
 		<div><a data-o-grid-colspan="4" href="/graph/engagement/articles-read">Articles read</a></div>
+		<div><a data-o-grid-colspan="4" href="/graph/engagement/referrers">Referrers</a></div>
 	</div>
 </div>
